Bugfix for the "Unusable startup URL" issue.
The issue appeared with the switch to go 1.5. When the server
started with the default configuration it would open the
URL "http://0.0.0.0" for IPv4 or "http://[::]" for IPv6. But
these URLs are not accepted by most browser (especially not on Windows).

So I added a detection for wildcard URLs and replace these wildcard URLs
with "localhost" instead. And if the URL is an IPv6 address I will
also replace the brackets to the startup URL looks like
"http://localhost" now.